convert object component
