+++ Engine/src/component/Component.h
@@ -157,25 +157,6 @@ struct EngineAPI NativeScriptComponent : public Component
         static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
 };
 
-
-struct EngineAPI ObjectComponent : public Component
-{
-	ObjectComponent() = default;
-	ObjectComponent(Entity e, const std::string& name) : name(name), e(e) {};
-
-	static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-
-	template <class Archive>
-	void serialize(Archive& archive) {
-		SERIALIZED_MEMBER(name);
-		SERIALIZED_MEMBER(e);
-	}
-
-    std::string name;
-    Entity e = Entity::EmptyEntity;
-    
-};
-
 struct EngineAPI InstanceBatch : public Component
 {
 public:
